requirements.txt
python-gitlab==3.15.0
gitpython==3.1.40
pandas==2.1.4
python-dotenv==1.0.0
